creating file_fun, gui_widgets, and swath_coverage_lib
transitioning swath_coverage_plotter.py into GUI setup that uses common file handling functions in file_fun.py and swath coverage-specific functions in swatch_coverage_lib.py, with the long-term goal of being able to share these across GUIs or plot without GUI
